Fix step idx in single-sbatch trajectory
